Reformat specification to improve readability
There should now be 1 sentence per line. This helps improve the
readability of intraline differences, and reduce the amount of lines
changes.

Move references to the bottom of sections, to reduce the strain required
in reading the Markdown version of the specification.

I also corrected a spelling error and linkified RFC2119.